+++ domain/model/src/main/java/com/crediya/model/constants/SalaryBaseLimits.java
@@ -1,10 +1,12 @@
 package com.crediya.model.constants;
 
+import java.math.BigDecimal;
+
 public class SalaryBaseLimits {
 
     private SalaryBaseLimits() {}
 	
-	public static final int MIN = 0;
-	public static final int MAX = 15_000_000;
+	public static final BigDecimal MIN = BigDecimal.valueOf(0);
+	public static final BigDecimal MAX = BigDecimal.valueOf(15_000_000);
 	
 }
