+++ infrastructure/entry-points/reactive-web/src/main/java/com/crediya/api/dto/input/user/CreateUserRequest.java
@@ -1,5 +1,6 @@
 package com.crediya.api.dto.input.user;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CreateUserRequest(
@@ -10,6 +11,6 @@ public record CreateUserRequest(
 	String direccion,
 	String telefono,
 	String correoElectronico,
-	Integer salarioBase
+	BigDecimal salarioBase
 	
 ) {}
